search-service/requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-dotenv==1.0.1
